Move a few setup methods back into pool
These are really pool methods, since they have to mess with filesystems,
and so forth. It is also an opportunity to make some StratPool methods
and constants private.

Make setup.rs comment more accurate.
