+++ src/engine/strat_engine/mod.rs
@@ -11,7 +11,7 @@ pub mod engine;
 pub mod metadata;
 mod mdv;
 pub mod filesystem;
-mod pool;
+pub mod pool;
 pub mod serde_structs;
 pub mod setup;
 mod range_alloc;
